requirements.txt
pandas
python-binance
asynctest
requests
pyodbc
PyYAML